Move DM functions to dao_setup
