+++ scripts_with_dao/dao_linearity_curve_3s_pyr.py
@@ -19,7 +19,7 @@ ROOT_DIR = config.root_dir
 
 # Import Specific Modules
 import dao
-from src.dao_setup import init_setup
+from src.dao_setup import init_setup, set_data_dm
 setup = init_setup()  # Import all variables from setup
 from src.utils import *
 from src.circular_pupil_functions import *
