+++ scripts_with_dao/dao_push-pull_calibration_3s_pyr.py
@@ -32,7 +32,7 @@ from src.utils import *
 from src.dao_create_flux_filtering_mask import *
 from src.psf_centring_algorithm import *
 from src.calibration_functions import *
-from src.dao_setup import init_setup
+from src.dao_setup import init_setup, set_dm_actuators
 setup = init_setup()  # Import all variables from setup
 
 #%% Accessing Devices
