+++ src/utils.py
@@ -56,102 +56,6 @@ def compute_data_slm(data_dm=0, data_phase_screen=0, data_dm_flat=0, setup=None,
 
     return data_slm.astype(np.uint8)
 
-
-def set_dm_actuators(actuators=None, dm_flat=None, setup=None, **kwargs):
-    """Set DM actuators and update the shared memory grid.
-
-    Parameters
-    ----------
-    actuators : array_like, optional
-        Actuator values to apply. Defaults to zeros.
-    dm_flat : array_like, optional
-        Flat map added to the actuators. Defaults to ``setup.dm_flat``.
-    setup : object, optional
-        DAO setup providing defaults for the deformable mirror.
-    kwargs : optional
-        ``dm`` – Deformable mirror instance. Defaults to ``setup.deformable_mirror``.
-    """
-
-    if setup is None:
-        if DEFAULT_SETUP is None:
-            raise ValueError("No setup provided and no default registered.")
-        setup = DEFAULT_SETUP
-
-    if actuators is None:
-        actuators = np.zeros(setup.nact ** 2)
-    if dm_flat is None:
-        dm_flat = setup.dm_flat
-
-    actuators = np.asarray(actuators)
-
-    deformable_mirror = kwargs.get("deformable_mirror", setup.deformable_mirror)
-    if deformable_mirror is None:
-        raise ValueError("Deformable mirror instance must be provided")
-
-    deformable_mirror.actuators = actuators + dm_flat
-
-    dm_act_shm = shm.dm_act_shm
-    dm_act_shm.set_data(np.asarray(deformable_mirror.actuators).astype(np.float64).reshape(setup.nact, setup.nact))
-    
-
-# ---------------------------------------------------------------------------
-def set_data_dm(actuators=None, *, setup=None, dm_flat=None, **kwargs):
-    """Flatten the DM, optionally apply ``actuators`` and show the resulting
-    phase on the SLM.
-
-    Parameters
-    ----------
-    actuators : array_like, optional
-        Actuator values to apply. Defaults to zeros.
-    dm_flat : array_like, optional
-        Flat map added to the actuators. Defaults to ``setup.dm_flat``.
-    setup : object, optional
-        DAO setup providing defaults for devices and dimensions.
-    kwargs : optional
-        ``dm``                 – DM instance. Defaults to ``setup.deformable_mirror``.
-        ``slm``                – SLM instance. Defaults to ``setup.slm``.
-        ``npix_small_pupil_grid`` – DM grid size. Defaults to ``setup.npix_small_pupil_grid``.
-        ``pupil_setup``        – Setup for ``compute_data_slm``. Defaults to ``setup.pupil_setup``.
-        ``wait_time``          – Delay after sending data. Defaults to ``setup.wait_time`` or ``0``.
-
-    Returns
-    -------
-    tuple of ndarray
-        ``(data_dm, data_slm)`` arrays that were sent to the SLM.
-    """
-
-    if setup is None:
-        if DEFAULT_SETUP is None:
-            raise ValueError("No setup provided and no default registered.")
-        setup = DEFAULT_SETUP
-
-    slm = kwargs.get("slm", setup.slm)
-    if slm is None:
-        raise ValueError("SLM instance must be provided")
-        
-    deformable_mirror = kwargs.get("deformable_mirror", setup.deformable_mirror)
-    if deformable_mirror is None:
-        raise ValueError("Deformable mirror instance must be provided")
-
-    npix_small_pupil_grid = kwargs.get("npix_small_pupil_grid", setup.npix_small_pupil_grid)
-    wait_time = kwargs.get("wait_time", setup.wait_time)
-    pupil_setup = kwargs.get("pupil_setup", setup.pupil_setup)
-
-    deformable_mirror.flatten()
-    set_dm_actuators(actuators, dm_flat=dm_flat, setup=setup)
-
-    data_dm = np.zeros((npix_small_pupil_grid, npix_small_pupil_grid), dtype=np.float32)
-    data_dm[:, :] = deformable_mirror.opd.shaped / 2
-
-    data_slm = compute_data_slm(data_dm=data_dm, setup=pupil_setup)
-    slm.set_data(data_slm)
-    time.sleep(wait_time)
-
-    return actuators, data_dm, data_slm
-
-
-
-
 def create_psf_mask(psf, crop_size=100, radius=50):
     """
     Creates a mask for the PSF image by cropping a region around the center,
